requirements.txt
flask>=2.0.0
pandas>=1.3.0
mlflow>=2.0.0
scikit-learn>=1.0.0
azure-ai-ml>=1.0.0
azure-identity>=1.0.0
python-dotenv>=0.19.0
numpy>=1.24.4
packaging>=21.0
joblib>=1.3.2
azure-storage-blob>=12.19.0
azureml-mlflow>=1.53.0